+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
       files: requirements-dev.txt
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.12.12
+  rev: v0.13.0
   hooks:
     - id: ruff
       args: ["--fix", "--show-fixes"]
@@ -26,7 +26,7 @@ repos:
     - id: pyproject-fmt
 
 - repo: https://github.com/woodruffw/zizmor-pre-commit
-  rev: v1.12.1
+  rev: v1.13.0
   hooks:
     - id: zizmor
