+++ atn2obis/templates/meta.xml.j2
@@ -17,7 +17,7 @@
     <files>
       <location>{{ emof_filename.split("\\")[-1] }}</location>
     </files>
-    <coreid index="0" />{% for emof_colname in emof_cols_list %}{% if emof_colname != 'eventID' %}    
+    <coreid index="0" />{% for emof_colname in emof_cols_list %}{% if emof_colname != 'eventID' %}
       <field index="{{ loop.index - 1}}" term="{% if emof_colname in('measurementTypeID', 'measurementValueID','measurementUnitID') %}http://rs.iobis.org/obis/terms/{% else %}http://rs.tdwg.org/dwc/terms/{% endif %}{{ emof_colname }}"/>{% endif %}{% endfor %}
   </extension>{% endif %}
 </archive>
