+++ tests/test_fecth_nc.py
@@ -1,5 +1,2 @@
-from atn_to_obis.atn_nc_to_dwc import recursive_wget
-
-
 def test_from_recursive_wget():
     pass
